[libutils] alloc_array: use long for alloc() return because the result, may be negative if it is an error

+++ libs/utils/alloc_array.h
@@ -140,7 +140,7 @@ namespace utils
         }
         size_t allocated_element_count() const { return _element_count; };
 
-        size_t alloc()
+        long alloc()
         {
 
             for (size_t i = 0; i < array_count; i++)
@@ -149,7 +149,7 @@ namespace utils
                 {
                     _status[i] = true;
                     _element_count++;
-                    return i;
+                    return (long)i;
                 };
             }
             printf("can't allocate a new entry in alloc_array\n");
